+++ YasEngineSoftware/FibonacciPointsGenerator.hpp
@@ -2,14 +2,12 @@
 #define FIBONACCIPOINTSGENERATOR_HPP
 #include"PointsGenerator.hpp"
 #include"PointsSet.hpp"
-#include"Math.hpp"
-#include"Vector2D.hpp"
 #include "YasGraphicsLibrary.hpp"
 
 class FibonacciPointsGenerator : public PointsGenerator
 {
 	public:
-        static std::vector<float> numbers;
+        static std::vector<int> numbers;
 		virtual PointsSet* generatePoints() override;
 
 };
