+++ YasEngineSoftware/PrimeNumbersPointsGenerator.cpp
@@ -1,13 +1,14 @@
 #include"PrimeNumbersPointsGenerator.hpp"
+#include "YasGraphicsLibrary.hpp"
 
-std::vector<float> PrimeNumbersPointsGenerator::numbers;
+std::vector<int> PrimeNumbersPointsGenerator::numbers;
 
 PointsSet* PrimeNumbersPointsGenerator::generatePoints()
 {
 	pointsNumber = 1000;
 
 	std::vector<int> primeNumbers = generatePrimeNumbersLessThanN(pointsNumber);
-    for(int i=0; i<primeNumbers.size(); i++)
+    for(int i=0; i<static_cast<int>(primeNumbers.size()); i++)
     {
         numbers.push_back((primeNumbers.at(i)));
     }
@@ -16,17 +17,12 @@ PointsSet* PrimeNumbersPointsGenerator::generatePoints()
 
 	Vector2D<float>* points = new Vector2D<float>[pointsNumber * 2];
 
-
-	// Point constructor setting x and y to 0;
-	// This loop creates pairs of points = ( (0,0) ; (fibonacci(i) ,y=0) )
 	int j = 0;
 	for (int i = 1; i < (pointsNumber * 2) - 1; i += 2)
 	{
 		points[i].x = static_cast<float>(primeNumbers.at(j++));//100.0F; // static_cast<float>(fibonacci.at(j++));
 	}
 
-	// DLUGOSC LINI JEST TEZ FIBONACCI NARAZIE
-	// ALE WIECEJ LINI BEDZIE JESLI DLUGOSCI ZROBIE STALE
 	j = 0;
 	for (int i = 1; i < pointsNumber * 2 - 1; i += 2)
 	{
